Refactor home components
Changed the name of the 'theme' folder to 'styles' to better
reflect the content.
