+++ frontend/src/components/views/home/styles/settings.js
